refactor(input): extract Enter/Tab handlers from handleKeypress (#1072)
Pull `handleReturnKey` and `handleTabKey` into a new
`reader.keypress.submit.ts` sibling, following the established
reader.keypress.{nav,paste} extraction pattern.

`handleKeypress` drops from 202 LOC to ~117 LOC, clearing the
200-line advisory function ceiling and removing its entry from
`.funcsize-baseline.json`.

No behavioral change — the external signature of `handleKeypress` is
unchanged and all callers are unaffected.

Closes #1072
